\name{AnnotationHubRecipes}
\docType{class}
\alias{AnnotationHubRecipes--class}

\title{Build and Run AnnotationHubRecipes}

\description{

}
  
\usage{
## Constructors
AnnotationHubRecipes(dcfFile, ...)
}
\arguments{

  \item{dcfFile}{The name of a DCF file}.
  \item{...}{As yet unused}
}


\keyword{classes}
