\name{EncodeBaseURL}
\alias{EncodeBaseURL}
\title{EncodeBaseURL}
\description{
The UCSC Genome Browser web site hosts Encode project data at this URL.
}
\usage{
EncodeBaseURL()
}
\value{
A character string following the standard HTTP format.
}
\author{Paul Shannon}

\keyword{utilities}
\keyword{IO}
