\name{flog}
\alias{flog}
\title{flog}

\description{

 Write logging message to console and a file.

}

\usage{
flog(level, ...)
}

\arguments{
  \item{level}{A \code{characater(1)} string object.}
  \item{\dots}{Further arguments.}
}

\details{

Writes the message to the console and to a file.
}

\value{
  None.
}

\author{Dan Tenenbaum}

\seealso{\code{futile.logger}}

%%\example{
%%   perhaps dan can provide one...
%% }

\keyword{classes}
