% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructHubFunctions.R
\name{hub_create_package}
\alias{hub_create_package}
\title{Function for creating a hub package}
\usage{
hub_create_package(package, type = c("AH", "EH"))
}
\arguments{
\item{package}{The name of the hub package to be created.}

\item{type}{Either \code{AH} or \code{EH} depending on if the package should 
be an AnnotationHub package or ExperimentHub package.}
}
\value{
Path to newly created package
}
\description{
Function for creating a hub package
}
\examples{
hub_create_package("test_hub", "AH")
}
