\name{validationFunctions}
\alias{validationFunctions}

\alias{getSpeciesList}
\alias{validSpecies}
\alias{suggestSpecies}
\alias{getValidSourceTypes}

\title{ValidationFunctions}

\description{
  Functions to assist in the validation process of creating the
  metadata.csv file for Hub Resources
}

\usage{
getSpeciesList(verbose=FALSE)

validSpecies(species, verbose=TRUE)

suggestSpecies(query, verbose=FALSE, op=c("|", "&"))

getValidSourceTypes()
}

\arguments{
  \item{species}{
    species to validate (may be single value or list)
  }
  \item{query}{
    terms to query. Whether AND or OR is determined by argument op.
  }
  \item{verbose}{
    should additional information and useful tips be displayed
  }
  \item{op}{
    Should searching of mulitple terms be conditional OR ("|") or AND
    ("&")
  }
}

\details{
   \itemize{
     \item getSpeciesList: Provides a list of valid species as
   determined by the GenomeInfoDbData package specData.rda file.

     \item validSpecies: True/False if argument is considered a valid
     species based on the list generated by getSpeciesList. A species
     may be deemed invalid if the capitalization mismatches or
     punctuation mismatches. Use suggestSpecies to find similar terms.

     \item suggestSpecies: Based on a term or multiple terms suggest
     possible valid species.

     \item getValidSourceTypes: returns list of acceptable values for
     SourceType in metadata.csv.  If you think a valid source type
     should be added to the list please reach out to
     maintainer@bioconductor.org


  }
}

\value{
  \itemize{
    \item For getSpeciesList: character vector of valid species
    \item For validSpecies: True/False if all species given as argument
    are valid
    \item For suggestSpecies: data.frame of taxonomy id and species name
    of possible valid species based on given query key words.
    \item For getValidSourceTypes: character vector of valid source types.
  }
}

\author{Lori Shepherd}

\seealso{
  \itemize{
    \item \link{AnnotationHubMetadata}
    \item \link{makeAnnotationHubMetadata}
  }
}

\examples{

species = getSpeciesList()

# following is TRUE

validSpecies("Homo sapiens")
# followin is FALSE because of starting "h"
validSpecies("homo sapiens")

# can provide multiple, if any are not valid FALSE
# TRUE
validSpecies(c("Homo sapiens", "Canis domesticus"))

suggestSpecies("Canis")

getValidSourceTypes()

}

\keyword{methods}
