\name{AnnotationHubMetadata}
\docType{class}
\alias{AnnotationHubMetadata--class}

\title{Build and Run AnnotationHubMetadata}

\description{

}
  
\usage{
## Constructors
AnnotationHubMetadata(dcfFile, ...)
}
\arguments{

  \item{dcfFile}{The name of a DCF file}.
  \item{...}{As yet unused}
}


\keyword{classes}
