\name{ensemblGtfToGRanges}
\alias{ensemblGtfToGRanges}
\title{ensemblGtfToGRanges}
\description{
Transform an Ensembl GTF file to a Bioconductor GRanges object.
}
\usage{
ensemblGtfToGRanges(recipe)
}
\arguments{
  \item{recipe}{a \code{AnnotationHubRecipe} object.}
}

\value{
None.
}
\author{Martin Morgan}

\seealso{
  extendedBedToGRanges,
  extendedBedWithAuxiliaryTableToGRanges
}


\keyword{classes}
