\name{extendedBedWithAuxiliaryTableToGRanges}
\alias{extendedBedWithAuxiliaryTableToGRanges}
\title{extendedBedWithAuxiliaryTableToGRanges}
\description{
Transform an 'extended' BED file (with possibly more than the canonical
three columns) into a Bioconductor GRanges object.
}
\usage{
extendedBedWithAuxiliaryTableToGRanges(recipe)
}
\arguments{
  \item{recipe}{a \code{AnnotationHubRecipe} object. }
}

\value{
None.
}
\author{Paul Shannon}

\seealso{
  extendedBedWithAuxiliaryTableToGRanges,
}


\keyword{classes}
