\name{speciesMap}
\alias{speciesMap}
\docType{data}
\title{
Species Mapper object
}
\description{
A data frame containing a mapping between NCBI Taxonomy IDs and 
species names.
}
\usage{data(speciesMap)}
\format{
  A data frame with 1416382 observations on the following 2 variables.
  \describe{
    \item{\code{taxon}}{a numeric vector}
    \item{\code{species}}{a character vector}
  }
}
\examples{
data(speciesMap)
}
\keyword{datasets}
